fix (Test) : 위치 변경
